Refactor: Migrate the outdated, unstable platform abstraction layer to Services
