+++ src/modules/essential/ModuleKeyValueDB.ts
@@ -5,6 +5,8 @@ import { LOG_LEVEL_NOTICE, LOG_LEVEL_VERBOSE } from "octagonal-wheels/common/log
 import { AbstractModule } from "../AbstractModule.ts";
 import type { LiveSyncCore } from "../../main.ts";
 import type { SimpleStore } from "octagonal-wheels/databases/SimpleStoreBase";
+import type { InjectableServiceHub } from "@/lib/src/services/InjectableServices.ts";
+import type { ObsidianDatabaseService } from "../services/ObsidianServices.ts";
 
 export class ModuleKeyValueDB extends AbstractModule {
     async tryCloseKvDB() {
@@ -77,6 +79,7 @@ export class ModuleKeyValueDB extends AbstractModule {
                     .filter((e) => e.startsWith(prefix))
                     .map((e) => e.substring(prefix.length));
             },
+            db: Promise.resolve(getDB()),
         } satisfies SimpleStore<T>;
     }
     _everyOnInitializeDatabase(db: LiveSyncLocalDB): Promise<boolean> {
@@ -100,12 +103,12 @@ export class ModuleKeyValueDB extends AbstractModule {
         }
         return true;
     }
-    onBindFunction(core: LiveSyncCore, services: typeof core.services): void {
+    onBindFunction(core: LiveSyncCore, services: InjectableServiceHub): void {
         services.databaseEvents.onUnloadDatabase.addHandler(this._onDBUnload.bind(this));
         services.databaseEvents.onCloseDatabase.addHandler(this._onDBClose.bind(this));
         services.databaseEvents.onDatabaseInitialisation.addHandler(this._everyOnInitializeDatabase.bind(this));
         services.databaseEvents.onResetDatabase.addHandler(this._everyOnResetDatabase.bind(this));
-        services.database.openSimpleStore.setHandler(this._getSimpleStore.bind(this));
+        (services.database as ObsidianDatabaseService).openSimpleStore.setHandler(this._getSimpleStore.bind(this));
         services.appLifecycle.onSettingLoaded.addHandler(this._everyOnloadAfterLoadSettings.bind(this));
     }
 }
